+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* [#41](https://github.com/fabiodomingues/clj-depend/issues/41): API to check if clj-depend is configured.
 * [#39](https://github.com/fabiodomingues/clj-depend/issues/39): Fix circular dependency not being treated as a violation.
 
 ## 0.8.1 (2023-09-04)